Move the old nodes into their own module
